fix: Prevent deadlock when releasing a realtime channel
The RealtimeChannels mutex is acquired to retrieve the channel relating
to an inbound message, so holding that mutex whilst waiting for Detach
to complete causes a deadlock because the inbound DETACHED message the
Detach call is waiting for also tries to acquire the mutex.

This commit fixes the deadlock by releasing the mutex before calling
Detach, and then acquiring it again once it completes in order to
release the channel.

Signed-off-by: Lewis Marshall <lewis.marshall@ably.com>